Tidy up (and document better) the code that allows specifying additional dependencies used as classpath for the plugin under test. This is needed for optional dependencies (as used by this plugin) that would normally be found through other applied plugins (testkit isolates the classpath, unlike regular gradle).
